requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
category_encoders==2.2.2
    # via -r requirements.in
joblib==1.0.1
    # via scikit-learn
numpy==1.21.2
    # via
    #   -r requirements.in
    #   category-encoders
    #   pandas
    #   patsy
    #   scikit-learn
    #   scipy
    #   statsmodels
pandas==1.3.2
    # via
    #   -r requirements.in
    #   category-encoders
    #   statsmodels
patsy==0.5.1
    # via
    #   category-encoders
    #   statsmodels
python-dateutil==2.8.2
    # via pandas
pytz==2021.1
    # via pandas
pyyaml==5.4.1
    # via -r requirements.in
scikit-learn @ git+https://github.com/thomasjpfan/scikit-learn.git@dd82abbf502d4a186b699a6e3d95fb06ce075de3
    # via
    #   -r requirements.in
    #   category-encoders
scipy==1.7.1
    # via
    #   -r requirements.in
    #   category-encoders
    #   scikit-learn
    #   statsmodels
six==1.16.0
    # via
    #   patsy
    #   python-dateutil
statsmodels==0.12.2
    # via category-encoders
threadpoolctl==2.2.0
    # via scikit-learn
